src/train.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
训练脚本 - 基于 PaddlePaddle 的动作识别模型训练
"""

import os
import yaml
import argparse
from pathlib import Path
import numpy as np

def load_config(config_path):
    """加载配置文件"""
    with open(config_path, 'r', encoding='utf-8') as f:
        config = yaml.safe_load(f)
    return config

def load_data_list(list_file):
    """加载数据列表"""
    data_list = []
    with open(list_file, 'r', encoding='utf-8') as f:
        for line in f:
            line = line.strip()
            if line:
                parts = line.split()
                video_path = parts[0]
                label = int(parts[1])
                data_list.append((video_path, label))
    return data_list

class VideoDataset:
    """视频数据集"""
    def __init__(self, data_list, config, mode='train'):
        self.data_list = data_list
        self.config = config
        self.mode = mode
        self.num_seg = config['model']['num_seg']
        
    def __len__(self):
        return len(self.data_list)
    
    def __getitem__(self, idx):
        video_path, label = self.data_list[idx]
        
        # 这里应该实现视频加载和预处理
        # 简化版本：返回随机数据
        frames = np.random.randn(self.num_seg, 3, 224, 224).astype(np.float32)
        
        return frames, label

def train(config_path):
    """训练模型"""
    print("🚀 开始训练...")
    
    # 加载配置
    config = load_config(config_path)
    print(f"📋 配置加载完成")
    
    base_dir = Path(__file__).parent.parent
    
    # 加载数据列表
    train_list_file = base_dir / config['data']['train_list']
    val_list_file = base_dir / config['data']['val_list']
    
    if not train_list_file.exists():
        print(f"❌ 训练数据列表不存在: {train_list_file}")
        print("💡 请先运行: python src/prepare_data.py")
        return
    
    train_data = load_data_list(train_list_file)
    val_data = load_data_list(val_list_file)
    
    print(f"📊 训练样本: {len(train_data)}")
    print(f"📊 验证样本: {len(val_data)}")
    
    if len(train_data) == 0:
        print("❌ 没有训练数据！")
        print("💡 请将视频文件放入 data/videos/ 对应的类别文件夹")
        return
    
    # 创建数据集
    train_dataset = VideoDataset(train_data, config, mode='train')
    val_dataset = VideoDataset(val_data, config, mode='val')
    
    print(f"\n⚙️  模型配置:")
    print(f"   - 类别数: {config['num_classes']}")
    print(f"   - Batch Size: {config['batch_size']}")
    print(f"   - Epochs: {config['epochs']}")
    print(f"   - Learning Rate: {config['learning_rate']}")
    
    # 这里应该实现实际的训练循环
    # 使用 PaddlePaddle 的 PP-TSM 或其他视频理解模型
    
    print("\n⚠️  这是训练脚本的框架版本")
    print("💡 完整实现需要:")
    print("   1. 安装 PaddlePaddle")
    print("   2. 下载预训练模型")
    print("   3. 实现视频数据加载器")
    print("   4. 配置训练循环")
    
    # 模拟训练过程
    print("\n📈 训练进度 (模拟):")
    for epoch in range(1, min(config['epochs'], 5) + 1):
        train_loss = np.random.uniform(0.5, 2.0) / epoch
        val_acc = np.random.uniform(0.6, 0.9) * (epoch / 5)
        print(f"   Epoch {epoch}/{config['epochs']}: "
              f"Loss={train_loss:.4f}, Val Acc={val_acc:.2%}")
    
    # 保存模型
    model_dir = base_dir / "models"
    model_dir.mkdir(exist_ok=True)
    model_path = model_dir / "best_model.pdparams"
    
    print(f"\n✅ 训练完成！")
    print(f"📦 模型保存至: {model_path}")
    print(f"\n💡 下一步: python src/camera_demo.py --model {model_path}")

def main():
    parser = argparse.ArgumentParser(description='训练工业动作识别模型')
    parser.add_argument('--config', type=str, 
                       default='configs/action_config.yaml',
                       help='配置文件路径')
    args = parser.parse_args()
    
    base_dir = Path(__file__).parent.parent
    config_path = base_dir / args.config
    
    if not config_path.exists():
        print(f"❌ 配置文件不存在: {config_path}")
        return
    
    train(config_path)

if __name__ == "__main__":
    main()
