requirements.txt
paddlepaddle>=2.5.0
paddlepaddle-gpu>=2.5.0  # 如果有GPU
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
